network: move SNTP functions to separate file
